requirements.txt
Flask==3.1.0
gunicorn==23.0.0
mutagen==1.47.0
Pillow==11.0.0
